requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.6
pydantic>=2.9.0
pydantic-settings>=2.6.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
torch>=2.5.0
flair>=0.14.0
sentence-transformers>=3.0.0
transformers>=4.46.0
faiss-cpu>=1.9.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pypdf>=5.0.0
pytesseract>=0.3.13
Pillow>=11.0.0
pdf2image>=1.16.0
opencv-python>=4.8.0
python-magic>=0.4.27
requests>=2.31.0
python-dotenv>=1.0.1
tiktoken>=0.8.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.1
